@@ -23,89 +23,89 @@
     }
   },
   "dependencies": {
-    "@angular-architects/module-federation": "^15.0.0",
-    "@angular/animations": "^15.2.7",
-    "@angular/cdk": "^15.2.7",
-    "@angular/common": "^15.2.7",
-    "@angular/compiler": "^15.2.7",
-    "@angular/core": "^15.2.7",
-    "@angular/elements": "^15.2.7",
-    "@angular/forms": "^15.2.7",
-    "@angular/platform-browser": "^15.2.7",
-    "@angular/platform-browser-dynamic": "^15.2.7",
-    "@angular/router": "^15.2.7",
-    "@commitlint/cli": "^19.4.0",
-    "@commitlint/config-conventional": "^19.2.2",
-    "@ngrx/component": "^15.4.0",
-    "@ngrx/effects": "^15.4.0",
-    "@ngrx/router-store": "^15.4.0",
-    "@ngrx/store": "^15.4.0",
-    "@ngrx/store-devtools": "^15.3.0",
-    "@ngx-translate/core": "^14.0.0",
-    "@ngx-translate/http-loader": "^7.0.0",
-    "@onecx/accelerator": "^4.42.1",
-    "@onecx/angular-accelerator": "^4.42.1",
-    "@onecx/angular-auth": "^4.42.1",
-    "@onecx/angular-integration-interface": "^4.42.1",
-    "@onecx/angular-remote-components": "^4.42.1",
-    "@onecx/angular-webcomponents": "^4.42.1",
-    "@onecx/integration-interface": "^4.42.1",
-    "@onecx/keycloak-auth": "^4.42.1",
+    "@angular-architects/module-federation": "^18.0.4",
+    "@angular/animations": "^18.2.3",
+    "@angular/cdk": "^18.2.3",
+    "@angular/common": "^18.2.3",
+    "@angular/compiler": "^18.2.3",
+    "@angular/core": "^18.2.3",
+    "@angular/elements": "^18.2.3",
+    "@angular/forms": "^18.2.3",
+    "@angular/platform-browser": "^18.2.3",
+    "@angular/platform-browser-dynamic": "^18.2.3",
+    "@angular/router": "^18.2.3",
+    "@commitlint/cli": "^19.4.1",
+    "@commitlint/config-conventional": "^19.4.1",
+    "@ngrx/component": "^18.0.2",
+    "@ngrx/effects": "^18.0.2",
+    "@ngrx/router-store": "^18.0.2",
+    "@ngrx/store": "^18.0.2",
+    "@ngrx/store-devtools": "^18.0.2",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
+    "@onecx/accelerator": "^5.10.1",
+    "@onecx/angular-accelerator": "^5.10.1",
+    "@onecx/angular-auth": "^5.10.1",
+    "@onecx/angular-integration-interface": "^5.10.1",
+    "@onecx/angular-remote-components": "^5.10.1",
+    "@onecx/angular-webcomponents": "^5.10.1",
+    "@onecx/integration-interface": "^5.10.1",
+    "@onecx/keycloak-auth": "^5.10.1",
     "@onecx/nx-plugin": "1.8.1",
-    "@onecx/portal-integration-angular": "^4.42.1",
-    "@onecx/portal-layout-styles": "^4.42.1",
+    "@onecx/portal-integration-angular": "^5.10.1",
+    "@onecx/portal-layout-styles": "^5.10.1",
     "@webcomponents/webcomponentsjs": "^2.8.0",
-    "keycloak-angular": "^13.1.0",
-    "primeflex": "^3.3.0",
-    "primeicons": "^6.0.1",
-    "primeng": "15.2.1",
-    "rxjs": "~7.8.0",
-    "tslib": "^2.3.0",
-    "zod": "^3.22.1",
-    "zone.js": "~0.12.0"
+    "keycloak-angular": "^16.0.1",
+    "primeflex": "^3.3.1",
+    "primeicons": "^7.0.0",
+    "primeng": "^17.18.9",
+    "rxjs": "~7.8.1",
+    "tslib": "^2.7.0",
+    "zod": "^3.23.8",
+    "zone.js": "~0.14.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.7",
-    "@angular-devkit/core": "^15.2.7",
-    "@angular-devkit/schematics": "^15.2.7",
-    "@angular-eslint/eslint-plugin": "~15.0.0",
-    "@angular-eslint/eslint-plugin-template": "~15.0.0",
-    "@angular-eslint/template-parser": "~15.0.0",
-    "@angular/cli": "^15.2.7",
-    "@angular/compiler-cli": "^15.2.7",
-    "@angular/language-service": "^15.2.7",
-    "@nx/angular": "17.3.2",
-    "@nx/eslint": "17.3.2",
-    "@nx/eslint-plugin": "17.3.2",
-    "@nx/jest": "17.3.2",
-    "@nx/js": "17.3.2",
-    "@nx/workspace": "17.3.0",
-    "@openapitools/openapi-generator-cli": "^2.5.2",
-    "@schematics/angular": "^15.2.7",
-    "@swc-node/register": "~1.6.7",
-    "@swc/core": "~1.3.85",
-    "@swc/helpers": "~0.5.2",
-    "@types/jest": "^29.4.0",
-    "@types/node": "18.16.9",
-    "@typescript-eslint/eslint-plugin": "^6.13.2",
-    "@typescript-eslint/parser": "^6.13.2",
-    "eslint": "~8.48.0",
-    "eslint-config-prettier": "^9.0.0",
+    "@angular-devkit/build-angular": "^18.2.3",
+    "@angular-devkit/core": "^18.2.3",
+    "@angular-devkit/schematics": "^18.2.3",
+    "@angular-eslint/eslint-plugin": "~18.3.0",
+    "@angular-eslint/eslint-plugin-template": "~18.3.0",
+    "@angular-eslint/template-parser": "~18.3.0",
+    "@angular/cli": "~18.2.3",
+    "@angular/compiler-cli": "^18.2.3",
+    "@angular/language-service": "^18.2.3",
+    "@nx/angular": "19.7.0",
+    "@nx/eslint": "19.7.0",
+    "@nx/eslint-plugin": "19.7.0",
+    "@nx/jest": "19.7.0",
+    "@nx/js": "19.7.0",
+    "@nx/workspace": "19.7.0",
+    "@openapitools/openapi-generator-cli": "^2.13.5",
+    "@schematics/angular": "^18.2.3",
+    "@swc-node/register": "1.10.9",
+    "@swc/core": "1.7.24",
+    "@swc/helpers": "0.5.13",
+    "@types/jest": "29.5.12",
+    "@types/node": "22.5.4",
+    "@typescript-eslint/eslint-plugin": "^7.17.0",
+    "@typescript-eslint/parser": "^7.17.0",
+    "eslint": "8.57.0",
+    "eslint-config-prettier": "^9.1.0",
     "husky": "^9.1.5",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "jest-preset-angular": "~14.2.0",
+    "jest": "29.7.0",
+    "jest-environment-jsdom": "29.7.0",
+    "jest-preset-angular": "14.2.2",
     "jest-sonar": "^0.2.16",
     "jest-sonar-reporter": "^2.0.0",
     "js-yaml": "^4.1.0",
-    "jsonc-eslint-parser": "^2.1.0",
-    "ngx-translate-testing": "^6.1.0",
-    "nx": "17.3.0",
-    "prettier": "^2.6.2",
+    "jsonc-eslint-parser": "^2.4.0",
+    "ngx-translate-testing": "^7.0.0",
+    "nx": "19.7.0",
+    "prettier": "^3.3.3",
     "sonar-scanner": "^3.1.0",
-    "sonarqube-scanner": "^4.2.1",
-    "ts-jest": "^29.1.0",
-    "ts-node": "10.9.1",
-    "typescript": "~4.9.4"
+    "sonarqube-scanner": "^4.2.2",
+    "ts-jest": "^29.2.5",
+    "ts-node": "10.9.2",
+    "typescript": "~5.5"
   }
 }
