@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common'
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core'
 import { FormsModule, ReactiveFormsModule } from '@angular/forms'
 import { RouterModule } from '@angular/router'
-import { LetModule } from '@ngrx/component'
+import { LetDirective } from '@ngrx/component'
 import { EffectsModule } from '@ngrx/effects'
 import { BookmarksCreateUpdateComponent } from './pages/bookmarks-search/dialogs/bookmarks-create-update/bookmarks-create-update.component'
 import { providePortalDialogService } from '@onecx/portal-integration-angular'
@@ -23,7 +23,7 @@ import { BookmarksSearchEffects } from './pages/bookmarks-search/bookmarks-searc
   imports: [
     CommonModule,
     SharedModule,
-    LetModule,
+    LetDirective,
     PortalCoreModule.forMicroFrontend(),
     RouterModule.forChild(addInitializeModuleGuard(routes)),
     FormsModule,
