@@ -1,6 +1,7 @@
 import { Injectable, SkipSelf } from '@angular/core'
 import { ActivatedRoute, Router } from '@angular/router'
-import { Actions, concatLatestFrom, createEffect, ofType } from '@ngrx/effects'
+import { Actions, createEffect, ofType } from '@ngrx/effects'
+import { concatLatestFrom } from '@ngrx/operators'
 import { Action, Store } from '@ngrx/store'
 import {
   AppStateService,
